Python Automation
It will help us to send the mail directly to the people with a single click(python <SCRIPT NAME.py>. We can manage any database system to keep record the details about sent emails along with the timestamp for further analysis. Here I am using to send the emails to the recruiters directly along with attachments as cover letter and resume.